refactor: deduplicate getExpiryStatus into shared utility (closes #3002)
Extract identical getExpiryStatus function from deliveries and products
route files into src/lib/expiry-utils.ts, replacing both local definitions
with a single import.
