+++ src/routes/_app/_auth/dashboard/_layout.gabinet.deliveries.tsx
@@ -1,5 +1,6 @@
 import { useState, useMemo, useCallback } from "react";
 import { createFileRoute } from "@tanstack/react-router";
+import { getExpiryStatus } from "@/lib/expiry-utils";
 import { useAction } from "convex/react";
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { api } from "@cvx/_generated/api";
@@ -123,20 +124,7 @@ function statusBadge(status: string, t: ReturnType<typeof useTranslation>["t"]) 
   );
 }
 
-function getExpiryStatus(dateStr: string | null): "expired" | "expiring_soon" | "ok" | null {
-  if (!dateStr) return null;
-  const today = new Date();
-  today.setHours(0, 0, 0, 0);
-  const parts = dateStr.split("-").map(Number);
-  if (parts.length !== 3) return null;
-  // Parse as local midnight so the expiry day itself is never flagged as expired
-  const expiry = new Date(parts[0], parts[1] - 1, parts[2]);
-  if (expiry < today) return "expired";
-  const soon = new Date(today);
-  soon.setDate(today.getDate() + 30);
-  if (expiry <= soon) return "expiring_soon";
-  return "ok";
-}
+
 
 function formatDate(ms: number | null | undefined) {
   if (!ms) return "—";
